# docs-site/openapi.yaml
openapi: 3.1.0
info:
  title: Data Gateway API
  version: 1.0.0
  description: API REST para administrar fuentes de datos, mappings y ejecutar consultas híbridas sin protocolo MCP.
servers:
  - url: https://data.whaapy.com
    description: Production API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
  schemas:
    NormalizedFilter:
      type: object
      required: [field, op, value]
      properties:
        field:
          type: string
        op:
          type: string
          enum: [eq, neq, gt, gte, lt, lte, in, contains, containsAny, containsAll]
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: array
              items:
                oneOf:
                  - type: string
                  - type: number
                  - type: boolean
    QueryPreference:
      allOf:
        - $ref: '#/components/schemas/NormalizedFilter'
        - type: object
          properties:
            boost:
              type: number
              minimum: 0
              maximum: 1
...
security:
  - bearerAuth: []
paths:
...
  /evals/sets:
    get:
      summary: Listar Evaluation Sets
      description: Devuelve los sets de evaluación del workspace. Requiere el scope `evals:read`.
      operationId: listEvalSets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      summary: Crear Evaluation Set
      description: Crea un set de evaluación global o asociado a una fuente. Requiere el scope `evals:write`.
      operationId: createEvalSet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                threshold:
                  type: number
                  minimum: 0
                  maximum: 1
                  default: 0.8
                sourceId:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Created
  /evals/sets/{id}:
    get:
      summary: Obtener Evaluation Set
      description: Devuelve un set de evaluación con todos sus cases. Requiere el scope `evals:read`.
      operationId: getEvalSet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
  /evals/sets/{id}/cases:
    post:
      summary: Crear Eval Case
      description: Agrega un caso de evaluación a un set. Requiere el scope `evals:write`.
      operationId: createEvalCase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                expectedExternalIds:
                  type: array
                  items:
                    type: string
                expectedTopIds:
                  type: array
                  items:
                    type: string
                mustRankAbove:
                  type: array
                  items:
                    type: object
                    required: [higher, lower]
                    properties:
                      higher:
                        type: string
                      lower:
                        type: string
                mustApplyFilters:
                  type: array
                  items:
                    $ref: '#/components/schemas/NormalizedFilter'
                mustApplyPreferences:
                  type: array
                  items:
                    $ref: '#/components/schemas/QueryPreference'
                mustNotContainFields:
                  type: array
                  items:
                    type: string
              anyOf:
                - required: [expectedExternalIds]
                - required: [expectedTopIds]
                - required: [mustRankAbove]
                - required: [mustApplyFilters]
                - required: [mustApplyPreferences]
                - required: [mustNotContainFields]
      responses:
        "201":
          description: Created
  /evals/sets/{id}/cases/{caseId}:
    delete:
      summary: Eliminar Eval Case
      description: Elimina un caso de evaluación mal creado antes de correr el set. Requiere el scope `evals:write`.
      operationId: deleteEvalCase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "404":
          description: Eval set or eval case not found
  /evals/run:
    post:
      summary: Correr Evaluation Set
      description: Encola una corrida de evaluación. Requiere el scope `evals:write`.
      operationId: runEvalSet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - evalSetId
              properties:
                evalSetId:
                  type: string
                  format: uuid
      responses:
        "202":
          description: Queued
  /evals/runs/{id}:
    get:
      summary: Obtener Eval Run
      description: Devuelve estado, métricas y resultados de una corrida. Requiere el scope `evals:read`.
      operationId: getEvalRun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK